requirements.txt
streamlit
pandas
PyPDF2
pdfminer.six
python-docx
scikit-learn
spacy
sentence-transformers
torch
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz